// src/model/ICompetitors.ts
export interface ICompetitors {
  ID: number;
  Category: string;
  Name: string;
  EngagementArea: string;
  TeamSize : number;
  BillingType: string;
  EngagementRate: string;
  PrimaryTechnology: string;
  SecondaryTechnology: string;
  AccountNameId: number;
  AccountName: any;
  Account_Name: string;
};